requirements.txt
pyserial==3.5
gpiozero==1.5.1
paho-mqtt==2.1.0
rpi.gpio==0.7.1